requirements.txt
asgiref==3.8.1
astroid==3.3.5
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.5.3
cryptography==42.0.8
defusedxml==0.8.0rc2
dill==0.3.9
Django==5.0.4
django-allauth==0.63.3
django-environ==0.11.2
django-storages==1.14.3
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pip-autoremove==0.10.0
platformdirs==4.3.6
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pylint==3.3.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==70.0.0
social-auth-core==4.5.4
sqlparse==0.5.0
tomlkit==0.13.2
tzdata==2024.1
urllib3==2.2.1
